+++ elphick/mass_composition/__init__.py
@@ -1,6 +1,8 @@
 from __future__ import annotations
 
 from .mass_composition import MassComposition
+from .stream import Stream
+from .flowsheet import Flowsheet
 from .mc_xarray import MassCompositionAccessor
 
 from importlib import metadata