+++ examples/106_network_basics.py
@@ -12,9 +12,8 @@ from typing import Dict
 import pandas as pd
 from matplotlib import pyplot as plt
 
-from elphick.mass_composition import MassComposition
+from elphick.mass_composition import MassComposition, Flowsheet
 from elphick.mass_composition.mc_node import MCNode
-from elphick.mass_composition.network import MCNetwork
 from elphick.mass_composition.datasets.sample_data import sample_data
 
 # %%
@@ -33,28 +32,28 @@ obj_mc_1, obj_mc_2 = obj_mc.split(0.4, name_1='stream 1', name_2='stream 2')
 # This is done to capture the relationships implicitly defined by any math operations performed on the objects.
 
 for obj in [obj_mc, obj_mc_1, obj_mc_2]:
-    print(obj.name, obj.nodes)
+    print(obj.name, obj._nodes)
 
 # %%
 #
-# Create a MCNetwork object
+# Create a Flowsheet object
 # -------------------------
 #
 # This requires passing an Iterable of MassComposition objects
 
-mcn: MCNetwork = MCNetwork().from_streams([obj_mc, obj_mc_1, obj_mc_2])
+fs: Flowsheet = Flowsheet().from_streams([obj_mc, obj_mc_1, obj_mc_2])
 
 # %%
 # Print the node object detail
 
-for node in mcn.graph.nodes:
-    print(mcn.graph.nodes[node]['mc'])
+for node in fs.graph.nodes:
+    print(fs.graph.nodes[node]['mc'])
 
 # %%
 # Note that the random node placeholder integers have been renumbered for readability.
 
 for obj in [obj_mc, obj_mc_1, obj_mc_2]:
-    print(obj.name, obj.nodes)
+    print(obj.name, obj._nodes)
 
 # %%
 # Print the overall network balanced status
@@ -62,47 +61,47 @@ for obj in [obj_mc, obj_mc_1, obj_mc_2]:
 # NOTE: presently this only includes node balance status
 # edge balance status will assure the mass-moisture balance is satisfied
 
-print(mcn.balanced)
+print(fs.balanced)
 
 # %%
 # Plot the network.
 # Imbalanced Nodes will appear red.  Later, Imbalanced Edges will also appear red.
 
-mcn.plot()
+fs.plot()
 plt
 
 # %%
 # Display the weight averages for all edges (streams) in the network (flowsheet)
 
-df_report: pd.DataFrame = mcn.report()
+df_report: pd.DataFrame = fs.report()
 df_report
 
 # %%
 
-df_report: pd.DataFrame = mcn.report(apply_formats=True)
+df_report: pd.DataFrame = fs.report(apply_formats=True)
 df_report
 
 # %%
 # Plot the interactive network using plotly
 
-fig = mcn.plot_network()
+fig = fs.plot_network()
 fig
 
 # %%
 # Plot the Sankey
 
-fig = mcn.plot_sankey()
+fig = fs.plot_sankey()
 fig
 
 # %%
 # Demonstrate the table-plot
 
-fig = mcn.table_plot(plot_type='sankey', table_pos='top', table_area=0.3)
+fig = fs.table_plot(plot_type='sankey', table_pos='top', table_area=0.3)
 fig
 
 # %%
 
-fig = mcn.table_plot(plot_type='network', table_pos='bottom', table_area=0.3)
+fig = fs.table_plot(plot_type='network', table_pos='bottom', table_area=0.3)
 fig
 
 # %%
@@ -114,9 +113,9 @@ fig
 obj_mc_3, obj_mc_4 = obj_mc_2.split(0.8, name_1='stream 3', name_2='stream 4')
 obj_mc_5 = obj_mc_1.add(obj_mc_3, name='stream 5')
 
-mcn2: MCNetwork = MCNetwork().from_streams([obj_mc, obj_mc_1, obj_mc_2, obj_mc_3, obj_mc_4, obj_mc_5])
+fs2: Flowsheet = Flowsheet().from_streams([obj_mc, obj_mc_1, obj_mc_2, obj_mc_3, obj_mc_4, obj_mc_5])
 
-fig = mcn2.table_plot(plot_type='sankey', table_pos='left')
+fig = fs2.table_plot(plot_type='sankey', table_pos='left')
 fig
 
 # %%
@@ -124,12 +123,12 @@ fig
 # Setting Node names
 # ------------------
 
-nodes_before: Dict[int, MCNode] = mcn.nodes_to_dict()
+nodes_before: Dict[int, MCNode] = fs.nodes_to_dict()
 print({n: o.node_name for n, o in nodes_before.items()})
 
 # %%
-mcn.set_node_names(node_names={0: 'node_0', 1: 'node_1', 2: 'node_2', 3: 'node_3'})
-nodes_after: Dict[int, MCNode] = mcn.nodes_to_dict()
+fs.set_node_names(node_names={0: 'node_0', 1: 'node_1', 2: 'node_2', 3: 'node_3'})
+nodes_after: Dict[int, MCNode] = fs.nodes_to_dict()
 print({n: o.node_name for n, o in nodes_after.items()})
 
 # %%
@@ -139,22 +138,22 @@ print({n: o.node_name for n, o in nodes_after.items()})
 #
 # First we show how to easily access the stream data as a dictionary
 
-stream_data: Dict[str, MassComposition] = mcn.streams_to_dict()
+stream_data: Dict[str, MassComposition] = fs.streams_to_dict()
 print(stream_data.keys())
 
 # %%
 # We will replace stream 2 with the same data as stream 1.
 
-new_stream: MassComposition = deepcopy(mcn.get_edge_by_name('stream 1'))
+new_stream: MassComposition = deepcopy(fs.get_edge_by_name('stream 1'))
 # we need to rename to avoid a creating a duplicate stream name
 new_stream.name = 'stream 1 copy'
-mcn.set_stream_data({'stream 2': new_stream})
-print(mcn.streams_to_dict().keys())
+fs.set_stream_data({'stream 2': new_stream})
+print(fs.streams_to_dict().keys())
 
 # %%
 # Of course the network is now unbalanced as highlighted in the Sankey
 
-fig = mcn.table_plot()
+fig = fs.table_plot()
 fig
 
 # %%
@@ -167,30 +166,30 @@ fig
 #
 # Let's break the links for the _stream 1_.
 
-mcn.reset_stream_nodes(stream="stream 1")
-fig = mcn.table_plot()
+fs.reset_stream_nodes(stream="stream 1")
+fig = fs.table_plot()
 fig
 
 # %%
 # We'll now break all remaining connections (we could have done this from the start).
 
-mcn.reset_stream_nodes()
-fig = mcn.table_plot()
+fs.reset_stream_nodes()
+fig = fs.table_plot()
 fig
 
 # %%
 # Now we'll create some linkages - of course they will be completely rubbish and not balance.
 
-mcn.set_stream_parent(stream="stream 1", parent="Feed")
-mcn.set_stream_child(stream="stream 1", child="stream 1 copy")
-fig = mcn.table_plot()
+fs.set_stream_parent(stream="stream 1", parent="Feed")
+fs.set_stream_child(stream="stream 1", child="stream 1 copy")
+fig = fs.table_plot()
 fig
 
 # %%
 # Perhaps less useful, but possible, we can build relationships by setting nodes directly.
 
-mcn.reset_stream_nodes()
-mcn.set_stream_nodes(stream="stream 1", nodes=(1, 2))
-mcn.set_stream_nodes(stream="stream 1 copy", nodes=(2, 3))
-fig = mcn.table_plot()
+fs.reset_stream_nodes()
+fs.set_stream_nodes(stream="stream 1", nodes=(1, 2))
+fs.set_stream_nodes(stream="stream 1 copy", nodes=(2, 3))
+fig = fs.table_plot()
 fig