+++ examples/113_partition_models.py
@@ -25,7 +25,7 @@ from scipy.interpolate import PchipInterpolator
 
 from elphick.mass_composition import MassComposition
 from elphick.mass_composition.datasets.sample_data import size_by_assay
-from elphick.mass_composition.network import MCNetwork
+from elphick.mass_composition.flowsheet import Flowsheet
 from elphick.mass_composition.utils.partition import napier_munn
 from elphick.mass_composition.utils.pd_utils import calculate_partition
 
@@ -55,9 +55,9 @@ part_cyclone = partial(napier_munn, d50=0.150, ep=0.1, dim='size')
 # Separate the object using the defined partitions.  UF = Underflow, OF = Overflow
 
 mc_uf, mc_of = mc_feed.apply_partition(definition=part_cyclone, name_1='underflow', name_2='overflow')
-mcn: MCNetwork = MCNetwork().from_streams([mc_feed, mc_uf, mc_of])
+fs: Flowsheet = Flowsheet().from_streams([mc_feed, mc_uf, mc_of])
 
-fig = mcn.table_plot(table_pos='left',
+fig = fs.table_plot(table_pos='left',
                      sankey_color_var='Fe', sankey_edge_colormap='copper_r', sankey_vmin=50, sankey_vmax=70)
 fig