+++ examples/502_dag.py
@@ -17,7 +17,7 @@ import plotly
 from elphick.mass_composition import MassComposition
 from elphick.mass_composition.dag import DAG
 from elphick.mass_composition.datasets.sample_data import sample_data
-from elphick.mass_composition.network import MCNetwork
+from elphick.mass_composition.flowsheet import Flowsheet
 
 logging.basicConfig(level=logging.INFO, format='%(asctime)s - %(name)s - %(levelname)s - %(message)s')
 logger = logging.getLogger(__name__)
@@ -57,14 +57,14 @@ dag.run({'feed_1': mc_sample,
 # fig = dag.plot()
 
 # %%
-# Create a MCNetwork object from the dag, enabling all the usual network plotting and analysis methods.
+# Create a Flowsheet object from the dag, enabling all the usual network plotting and analysis methods.
 
-mcn: MCNetwork = MCNetwork.from_dag(dag)
+fs: Flowsheet = Flowsheet.from_dag(dag)
 
-fig = mcn.plot_network()
+fig = fs.plot_network()
 plotly.io.show(fig)
 
 # %%
 
-fig = mcn.table_plot(plot_type='sankey')
+fig = fs.table_plot(plot_type='sankey')
 fig