+++ tests/test_100_flowsheet.py
@@ -7,39 +7,38 @@ import pandas as pd
 import pytest
 
 from dev.debugging_tools import pretty_print_graph
-from elphick.mass_composition import MassComposition
+from elphick.mass_composition import MassComposition, Flowsheet
 from elphick.mass_composition.mc_node import MCNode
-from elphick.mass_composition.network import MCNetwork
 # noinspection PyUnresolvedReferences
-from tests.fixtures import demo_data, size_assay_data, demo_size_network, script_loc
+from .fixtures import demo_data, size_assay_data, demo_size_network, script_loc
 
 
 def test_graph_object_types_and_attributes(demo_size_network):
-    mcn: MCNetwork = demo_size_network
+    fs: Flowsheet = demo_size_network
 
-    pretty_print_graph(mcn.graph)
+    pretty_print_graph(fs.graph)
 
     # Check that all nodes are of type MCNode and have the 'mc' attribute
-    for node in mcn.graph.nodes:
+    for node in fs.graph.nodes:
         assert isinstance(node, (int, str))
-        assert 'mc' in mcn.graph.nodes[node]
-        assert isinstance(mcn.graph.nodes[node]['mc'], MCNode)
+        assert 'mc' in fs.graph.nodes[node]
+        assert isinstance(fs.graph.nodes[node]['mc'], MCNode)
 
     # Check that all edges are of type MassComposition and have the 'mc' attribute
-    for u, v, d in mcn.graph.edges(data=True):
+    for u, v, d in fs.graph.edges(data=True):
         assert isinstance(d['mc'], MassComposition)
 
 
 def test_sankey_plot(demo_size_network):
-    mcn: MCNetwork = demo_size_network
+    fs: Flowsheet = demo_size_network
     # test both types of colormaps
-    fig = mcn.plot_sankey(color_var='Fe', edge_colormap='copper_r', vmin=50, vmax=70)
-    fig2 = mcn.plot_sankey(color_var='Fe', edge_colormap='viridis')
+    fig = fs.plot_sankey(color_var='Fe', edge_colormap='copper_r', vmin=50, vmax=70)
+    fig2 = fs.plot_sankey(color_var='Fe', edge_colormap='viridis')
 
 
 def test_network_plot(demo_size_network):
-    mcn: MCNetwork = demo_size_network
-    fig = mcn.plot_network()
+    fs: Flowsheet = demo_size_network
+    fig = fs.plot_network()
     fig.show()
     pass
 
@@ -48,19 +47,19 @@ def test_table_plot(demo_data):
     obj_mc: MassComposition = MassComposition(demo_data, name='Feed')
     obj_mc_1, obj_mc_2 = obj_mc.split(0.4)
 
-    mcn: MCNetwork = MCNetwork().from_streams([obj_mc, obj_mc_1, obj_mc_2])
-    fig = mcn.table_plot()
-    fig = mcn.table_plot(plot_type='network', table_pos='right', table_area=0.3)
-    fig = mcn.table_plot(plot_type='network', table_pos='top', table_area=0.3)
-    fig = mcn.table_plot(table_pos='bottom', table_area=0.2)
+    fs: Flowsheet = Flowsheet().from_streams([obj_mc, obj_mc_1, obj_mc_2])
+    fig = fs.table_plot()
+    fig = fs.table_plot(plot_type='network', table_pos='right', table_area=0.3)
+    fig = fs.table_plot(plot_type='network', table_pos='top', table_area=0.3)
+    fig = fs.table_plot(table_pos='bottom', table_area=0.2)
 
 
 def test_to_dataframe(demo_data):
     obj_mc: MassComposition = MassComposition(demo_data, name='Feed')
     obj_mc_1, obj_mc_2 = obj_mc.split(0.4)
 
-    mcn: MCNetwork = MCNetwork().from_streams([obj_mc, obj_mc_1, obj_mc_2])
-    df_res: pd.DataFrame = mcn.to_dataframe()
+    fs: Flowsheet = Flowsheet().from_streams([obj_mc, obj_mc_1, obj_mc_2])
+    df_res: pd.DataFrame = fs.to_dataframe()
 
     d_expected: Dict = {'mass_wet': {(0, 'Feed'): 100.0, (1, 'Feed'): 90.0, (2, 'Feed'): 110.0,
                                      (0, '(0.4 * Feed)'): 40.0, (1, '(0.4 * Feed)'): 36.0,
@@ -101,12 +100,12 @@ def test_from_dataframe_tall(demo_data):
     logging.basicConfig(level=logging.DEBUG)
     obj_mc: MassComposition = MassComposition(demo_data, name='Feed')
     obj_mc_1, obj_mc_2 = obj_mc.split(0.4)
-    mcn: MCNetwork = MCNetwork().from_streams([obj_mc, obj_mc_1, obj_mc_2])
-    df_tall: pd.DataFrame = mcn.to_dataframe()
+    fs: Flowsheet = Flowsheet().from_streams([obj_mc, obj_mc_1, obj_mc_2])
+    df_tall: pd.DataFrame = fs.to_dataframe()
 
-    mcn_res: MCNetwork = MCNetwork().from_dataframe(df=df_tall,
+    fs_res: Flowsheet = Flowsheet().from_dataframe(df=df_tall,
                                                     mc_name_col='name')
-    df_res: pd.DataFrame = mcn_res.to_dataframe()
+    df_res: pd.DataFrame = fs_res.to_dataframe()
 
     pd.testing.assert_frame_equal(df_tall, df_res)
 
@@ -114,15 +113,15 @@ def test_from_dataframe_tall(demo_data):
 def test_from_dataframe_wide(demo_data):
     obj_mc: MassComposition = MassComposition(demo_data, name='Feed')
     obj_mc_1, obj_mc_2 = obj_mc.split(0.4)
-    mcn: MCNetwork = MCNetwork().from_streams([obj_mc, obj_mc_1, obj_mc_2])
-    df_wide: pd.DataFrame = mcn.to_dataframe().unstack('name')
+    fs: Flowsheet = Flowsheet().from_streams([obj_mc, obj_mc_1, obj_mc_2])
+    df_wide: pd.DataFrame = fs.to_dataframe().unstack('name')
     df_wide.columns = df_wide.columns.swaplevel(0, 1)
     df_wide.columns = ['_'.join(col) for col in df_wide.columns.values]
 
-    mcn_res: MCNetwork = MCNetwork().from_dataframe(df=df_wide, mc_name_col=None)
+    fs_res: Flowsheet = Flowsheet().from_dataframe(df=df_wide, mc_name_col=None)
 
-    df_test: pd.DataFrame = mcn.to_dataframe()
-    df_res: pd.DataFrame = mcn_res.to_dataframe()
+    df_test: pd.DataFrame = fs.to_dataframe()
+    df_res: pd.DataFrame = fs_res.to_dataframe()
     # indexes are in a different order...  So be it.
     assert set(df_res.index) == set(df_test.index)
     df_res = df_res.loc[df_test.index, :]
@@ -131,44 +130,44 @@ def test_from_dataframe_wide(demo_data):
 
 
 def test_from_yaml(script_loc):
-    mcn: MCNetwork = MCNetwork().from_yaml(flowsheet_file=Path(script_loc / 'config/flowsheet_example.yaml'))
+    fs: Flowsheet = Flowsheet().from_yaml(flowsheet_file=Path(script_loc / 'config/flowsheet_example.yaml'))
     with pytest.raises(KeyError):
-        mcn.report()
+        fs.report()
     with pytest.raises(KeyError):
-        mcn.plot_sankey()
+        fs.plot_sankey()
 
 
 def test_streams_to_dict(demo_size_network):
-    mcn: MCNetwork = demo_size_network
-    streams: Dict[str, MassComposition] = mcn.streams_to_dict()
+    fs: Flowsheet = demo_size_network
+    streams: Dict[str, MassComposition] = fs.streams_to_dict()
     for k, v in streams.items():
         assert isinstance(v, MassComposition)
         assert k == v.name
 
 
 def test_nodes_to_dict(demo_size_network):
-    mcn: MCNetwork = demo_size_network
-    nodes: Dict[int: MCNode] = mcn.nodes_to_dict()
+    fs: Flowsheet = demo_size_network
+    nodes: Dict[int: MCNode] = fs.nodes_to_dict()
     assert list(nodes.keys()) == [0, 1, 2, 3]
 
 
 def test_set_node_names(demo_size_network):
-    mcn: MCNetwork = demo_size_network
-    mcn.set_node_names(node_names={0: 'new_feed_name'})
-    assert mcn.graph.nodes[0]['mc'].node_name == 'new_feed_name'
+    fs: Flowsheet = demo_size_network
+    fs.set_node_names(node_names={0: 'new_feed_name'})
+    assert fs.graph.nodes[0]['mc'].node_name == 'new_feed_name'
 
 
 def test_set_stream_data(demo_size_network):
-    mcn: MCNetwork = demo_size_network
-    streams_original: Dict = deepcopy(mcn.streams_to_dict())
-    coarse: MassComposition = deepcopy(mcn.get_edge_by_name('coarse'))
+    fs: Flowsheet = demo_size_network
+    streams_original: Dict = deepcopy(fs.streams_to_dict())
+    coarse: MassComposition = deepcopy(fs.get_edge_by_name('coarse'))
     coarse.name = 'coarse_2'
-    mcn.set_stream_data(stream_data={'fine': coarse})
-    streams_modified: Dict = mcn.streams_to_dict()
+    fs.set_stream_data(stream_data={'fine': coarse})
+    streams_modified: Dict = fs.streams_to_dict()
 
     assert list(streams_modified.keys()) == ['size sample', 'coarse', 'coarse_2']
-    df1: pd.DataFrame = mcn.get_edge_by_name('coarse').data.to_dataframe()
-    df2: pd.DataFrame = mcn.get_edge_by_name('coarse_2').data.to_dataframe()
+    df1: pd.DataFrame = fs.get_edge_by_name('coarse').data.to_dataframe()
+    df2: pd.DataFrame = fs.get_edge_by_name('coarse_2').data.to_dataframe()
     pd.testing.assert_frame_equal(df1, df2)
 
 
@@ -177,11 +176,11 @@ def test_to_json(script_loc):
     pass
     # jsonpickle_numpy.register_handlers()
     # jsonpickle_pandas.register_handlers()
-    # mcn: MCNetwork = MCNetwork().from_yaml(flowsheet_file=Path(script_loc / 'config/flowsheet_example.yaml'))
-    # json_graph: Dict = mcn.to_json()
+    # fs: Flowsheet = Flowsheet().from_yaml(flowsheet_file=Path(script_loc / 'config/flowsheet_example.yaml'))
+    # json_graph: Dict = fs.to_json()
     # pickled_obj = jsonpickle.encode(json_graph)
     # unpickled_obj = jsonpickle.decode(pickled_obj)
-    # mcn2: MCNetwork = cytoscape_graph(unpickled_obj)
+    # fs2: Flowsheet = cytoscape_graph(unpickled_obj)
     #
     # with open('test_graph.json', 'w') as f:
     #     json.dump(pickled_obj, f)